src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-pseudo-element-with-host-pseudo-class-2-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Test for crbug.com/274063: cannot style ::content with a rule that includes :host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
